Article Likes 기존 복잡한 쿼리에서 간단한 jpa 메서드를 이용해서 로직 리팩토링

+++ dadaiksan/src/main/java/com/eanswer/dadaiksan/Dto/ArticleDto.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class ArticleDto {
     private Long id;
     private int viewCount = 0;
-    private BigInteger likeCounts;
+    private Long likeCounts;
     private String articleType;
     private String title;
     private String contents;
